Gillepsie algorithm

Mathematical foundations

The original article by Gillepsie:

A general method for numerically simulating the stochastic time evolution of coupled chemical reactions

Update Rates

$b_i$ and $d_i$ represent respcetively birth and death rates of agents $i$. The total rate is given by the sum of all individual rates

\[R(t) = \left[ \sum_i b_i(t) + d_i(t) \right]\]

A particular event, birth or death, is chosen at random with a probability equal to the rate of this event divided by the total rate $R$.

ABMEv.Gillepsie — Type

struct Gillepsie <: ABMEv.AbstractAlg

ABMEv.updateWorld! — Method

function updateWorld!(w::World{A,S,T},g::G,b,d)

Updating rule for gillepsie setting. Returning dt drawn from an exponential distribution with parameter the total rates of events.

ABMEv.update_rates! — Method

update_rates!(w::World,::Gillepsie,b,d)

This standard updates takes - competition kernels of the form α(x,y) and - carrying capacity of the form K(x)