Run the World

ABMEv.run! — Function

function run!(w::World{A,S,T},alg::L,tend::Number,b,d;dt_saving=nothing,cb=(names = String[],agg =nothing))

Run w with algorithm alg, until tend is reached. User needs to provide b the birth function, which takes as arguments X,t, and provide d the death function, with arguments X,Y,t. Returns a Simulation type.

if dt_saving specified, world is saved every time steps.

If dt_saving not specified, sim contains an array of two elements, first corresponding to initial conditions and last corresponding to world in the last time step.

if t_saving_cb::Array{Float64} specified, callbacks are computed at each steps time specified in the array.

This functionality is as of now only compatible with dt_saving not specified.

cb correspond to callbacks function. Look at the documentation for more information

the run stops if the number of agents reachesp["NMax"].

For now three algorithms

Missing docstring.

Missing docstring for Gillepsie. Check Documenter's build log for details.

Missing docstring.

Missing docstring for CFM. Check Documenter's build log for details.

ABMEv.WF — Type

struct WF <: ABMEv.AbstractAlg

Warning : WF not implemented yet