Space

EvoId.ContinuousSegment — Type

struct ContinuousSegment{T<:AbstractFloat} <: EvoId.AbstractSegment{T<:AbstractFloat}

Creates a segment space, where individuals are reflected at both ends.

Arguments

s start of the segment

e end of the segment

Example

ContinuousSegment(1., 2.)

EvoId.DiscreteSegment — Type

struct DiscreteSegment{T<:Integer} <: EvoId.AbstractSegment{T<:Integer}

Creates a discrete segement space, where individuals are reflected at both ends.

Arguments

s start of the segment

e end of the segment

Example

    DiscreteSegment(1, 2)

EvoId.GraphSpace — Type

struct GraphSpace{T} <: EvoId.AbstractStatSpace{1,T,EvoId.IsFinite{true}}

Creates a Graph Space.

Example

using LightGraphs
g = star_graph(7)
GraphSpace(g)

EvoId.NaturalSpace — Type

struct NaturalSpace{N, T} <: EvoId.AbstractStatSpace{N,T,EvoId.IsFinite{false}}

A natural space with dimension N and type T

EvoId.RealSpace — Type

struct RealSpace{N, T} <: EvoId.AbstractStatSpace{N,T,EvoId.IsFinite{false}}

Creates a real space.

Arguments

N dimension of the space

T type of the underlying traits.

EvoId.get_graph — Method

get_graph(d, t)

Returns the graph correseponding to d::DynGraphSpace at time t

EvoId.get_inc — Method

get_inc(x, D, s, t)

Returns increment corresponding to space s