Utils

ABMEv.DiversityFunction — Type

struct DiversityFunction

ABMEv.arithsmooth — Method

function arithsmooth(x,y,smooth)

-x is the shifted x-axis vector, due to smoothing -y is the smoothed value Returning a tuple

ABMEv.arithsmooth — Method

function arithsmooth(x,smooth)

arithmetic smoothing

ABMEv.arithsmooth2D — Method

function arithsmooth2D(xa,ya,itp,smooth)

Return xas,yas,A where A is an interpolated matrix with arithmetic smooth, whose axis are xas, yas

ARGS

xa xaxis values, ya yaxis values, itp interpolation function, smooth smoothing function

ABMEv.gaussian — Method

    gaussian(x::Number,mu::Number,sigma::Number) = generalised_gaussian(x,mu,sigma,2)

ABMEv.generalised_gaussian — Method

    generalised_gaussian(x::Number,mu::Number,sigma::Number,epsilon::Number)

Returns exp( -.5 * ((x-mu) / sigma)^epsilon)

ABMEv.geomsmooth — Method

function geomsmooth(x,y,smooth)

-x is the shifted x-axis vector, due to smoothing -y is the smoothed value Returning a tuple

ABMEv.geomsmooth — Method

function geomsmooth(x,smooth)

Geometric smoothing, cf https://en.wikipedia.org/wiki/Exponential_smoothing

ABMEv.geomsmooth2D — Method

function geomsmooth2D(xa,ya,itp,smooth)

Return xas,yas,A where A is an interpolated matrix with geometric smooth, whose axis are xas, yas

ARGS

xa xaxis values, ya yaxis values, itp interpolation function, smooth smoothing function

ABMEv.interpolate_df — Method

function interpolate_df(df,xlab,ylab,zlab)

returns an interpolated function itp(x,y) -> z, as well as its axis xa and ya

ABMEv.ma — Method

ma(x::Array{T},f) where T <: Number

Moving average over array x, using f as the filter, i.e. the number of points to average on. Better choosing an odd number

DataFrames.groupby — Method

function groupby(f, list::Array)

returns a dictionary that group list elements by value of function f