Agent

EvoId.Agent — Type

mutable struct Agent{A<:EvoId.Ancestors, R<:EvoId.Rates, T<:Tuple, U, V} <: AbstractAgent{A<:EvoId.Ancestors,R<:EvoId.Rates}

Specificities

The type Agent has two important composite types

Ancestors{bool} : when bool = true, the ancestors traits are stored,

Rates{bool} : when bool = true, the rates d and b of agents are updated at each time step. This is needed in e.g. Gillepsie algorithm.

EvoId.Agent — Method

Agent(s; ancestors=false,rates=true)
Agent(s, pos;ancestors=false,rates=true)

Returns an Agent living on the underlying space s with initial position pos. If pos not provided, initialises agent with 0 values everywhere

Keyword arguments

rates. Set rates=true when agents fitness needs to be updated at each time step. This

is required for the Gillepsie algorithm, but not for CFM algorithm

ancestors. Set ancestors=true when you want to store ancestors traits.

EvoId.get_geo — Method

get_geo(a, t)

Returns geotrait of agent a at time t.

EvoId.get_t — Method

get_t(a)

Get time when agent born.

EvoId.get_x — Method

get_x(a, t, i)

Returns trait i of the agent. Geotrait corresponds to dimension i=0.

EvoId.get_x — Method

get_x(a::Agent)

Returns trait i of the agent.

EvoId.increment_x! — Method

increment_x!(a, s, p, t)

This function increments agent by random numbers specified in p ONLY FOR CONTINUOUS DOMAINS