Run the World

ABMEv.run! — Function

function run!(w::World{A,S,T},alg::L,tend::Number;dt_saving=nothing,cb=(names = String[],agg =nothing))

Run w with algorithm alg, until tend is reached. Returns a Simulation type.

worldall stores the world every p["dt_saving"] time steps.

If dt_saving not specified, sim contains an array of two elements, first corresponding to initial conditions and last corresponding to world in the last time step.

the run stops if the number of agents reachesp["NMax"].

:warning: if you choose nagents = 1 then nothing is saved until the end of the simulation.

For now three algorithms

Missing docstring.

Missing docstring for Gillepsie. Check Documenter's build log for details.

Missing docstring.

Missing docstring for CFM. Check Documenter's build log for details.

ABMEv.WF — Type

struct WF <: ABMEv.AbstractAlg

Warning : WF not implemented yet