Space

ABMEv.AbstractSpace — Type

abstract type AbstractSpace

Dim is the dimension of the space, T is the element type, I to indicate finiteness

ABMEv.ContinuousSegment — Type

struct ContinuousSegment{T<:AbstractFloat} <: ABMEv.AbstractSegment{T<:AbstractFloat}

ABMEv.DiscreteSegment — Type

struct DiscreteSegment{T<:Integer} <: ABMEv.AbstractSegment{T<:Integer}

ABMEv.GraphSpace — Type

struct GraphSpace{T} <: ABMEv.AbstractStatSpace{1,T,ABMEv.IsFinite{true}}

ABMEv.NaturalSpace — Type

struct NaturalSpace{N, T} <: ABMEv.AbstractStatSpace{N,T,ABMEv.IsFinite{false}}

A natural space with dimension N and type T

ABMEv.RealSpace — Type

struct RealSpace{N, T} <: ABMEv.AbstractStatSpace{N,T,ABMEv.IsFinite{false}}

A real space with dimension N and type T

ABMEv._reflect1D — Method

_reflect1D(x, inc, s)

Here we increment the trajectory of trait 1 such that it follows a reflected brownian motion (1D)

ABMEv.get_graph — Method

get_graph(d, t)

Returns the graph correseponding to d::DynGraphSpace at time t

ABMEv.get_inc — Method

get_inc(x, D, s, t)

Returns increment corresponding to space s