Utils

DataFrames.groupby — Method

groupby(f, list::Array)

returns a dictionary that group list elements by value of function f

EvoId.arithsmooth — Method

arithsmooth(x,y,smooth)

Returning a tuple

Arguments

-x is the shifted x-axis vector, due to smoothing -y is the smoothed value

EvoId.arithsmooth — Method

arithsmooth(x,smooth)

arithmetic smoothing

EvoId.arithsmooth2D — Method

arithsmooth2D(xa,ya,itp,smooth)

Returns xas, yas, A where A is an interpolated matrix with arithmetic smooth, whose axis are xas, yas

Arguments

xa xaxis values, ya yaxis values, itp interpolation function, smooth smoothing function

EvoId.gaussian — Method

gaussian(x::Number,mu::Number,sigma::Number) = generalised_gaussian(x,mu,sigma,2)

EvoId.generalised_gaussian — Method

    generalised_gaussian(x::Number,mu::Number,sigma::Number,epsilon::Number)

Returns exp( -.5 * ((x-mu) / sigma)^epsilon)

EvoId.geomsmooth — Method

geomsmooth(x,y,smooth)

Returning a tuple

Arguments

-x is the shifted x-axis vector, due to smoothing -y is the smoothed value

EvoId.geomsmooth — Method

geomsmooth(x,smooth)

Geometric smoothing, cf https://en.wikipedia.org/wiki/Exponential_smoothing

EvoId.geomsmooth2D — Method

geomsmooth2D(xa,ya,itp,smooth)

Returns xas, yas, A where A is an interpolated matrix with geometric smooth, whose axis are xas, yas.

Arguments

xa xaxis values, ya yaxis values, itp interpolation function, smooth smoothing function

EvoId.interpolate_df — Method

interpolate_df(df,xlab,ylab,zlab)

returns an interpolated function itp(x,y) -> z, as well as its axis xa and ya

EvoId.ma — Method

ma(x::Array{T},f) where T <: Number

Moving average over array x, using f as the filter, i.e. the number of points to average on. Better choosing an odd number