World

ABMEv.get_geo — Method

get_geo(w)

ABMEv.get_x — Method

get_x(w, trait)

Returns trait of every agents of world in the form of an array which dimensions corresponds to the input. If trait = 0 , we return the geotrait.

Warning

Geotrait might be deprecated in the future.

ABMEv.get_xarray — Function

get_xarray(world)
get_xarray(world, geotrait)

Returns every traits of every agents of world in the form of a one dimensional array (in contrast to get_x). If geotrait=true the geotrait is also added to the set of trait, in the last column. If you do not want to specify t (only useful for geotrait), it is also possible to use get_xarray(world::Array{T,1}) where {T <: Agent}.

Warning

It does not work with subspace where ndims(subspace) > 1.

Base.getindex — Method

getindex(w, i)

Get x of world without geotrait.